# contracts/openapi.json
{
  "components": {
    "responses": {
      "ErrorResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Stable machine-readable error envelope."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
      "GroupCreateRequest": {
        "additionalProperties": false,
        "description": "The account-scoped group creation command.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "GroupCreateRequest",
        "type": "object"
      },
      "GroupResponse": {
        "additionalProperties": false,
        "description": "Server-owned group identity and settlement policy.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_account_id": {
            "title": "Owner Account Id",
            "type": "string"
          },
          "settlementPolicy": {
            "enum": [
              "owner_only",
              "any_member"
            ],
            "title": "Settlementpolicy",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "owner_account_id",
          "settlementPolicy"
        ],
        "title": "GroupResponse",
        "type": "object"
      },
      "GroupSummaryResponse": {
        "additionalProperties": false,
        "description": "Server-derived account-scoped group summary without child rows.",
        "properties": {
          "expenses_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expenses Count"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "member_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Count"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outings_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outings Count"
          },
          "owner_account_id": {
            "title": "Owner Account Id",
            "type": "string"
          },
          "participants_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participants Count"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "settlementPolicy": {
            "enum": [
              "owner_only",
              "any_member"
            ],
            "title": "Settlementpolicy",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "owner_account_id",
          "settlementPolicy",
          "role"
        ],
        "title": "GroupSummaryResponse",
        "type": "object"
      },
      "GroupUpdateRequest": {
        "additionalProperties": false,
        "description": "The only mutable group setting exposed by the MVP.",
        "properties": {
          "settlementPolicy": {
            "enum": [
              "owner_only",
              "any_member"
            ],
            "title": "Settlementpolicy",
            "type": "string"
          }
        },
        "required": [
          "settlementPolicy"
        ],
        "title": "GroupUpdateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cc_session": {
        "description": "Opaque server-recognized session cookie.",
        "in": "cookie",
        "name": "cc_session",
        "type": "apiKey"
      },
      "csrf_token": {
        "description": "Required on unsafe cookie-authenticated operations.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Cuentas Claras API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "description": "List only groups belonging to the authenticated account.",
        "operationId": "list_groups_api_v1_groups_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GroupSummaryResponse"
                  },
                  "title": "Response List Groups Api V1 Groups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Groups",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "description": "Create an empty owner workspace for the authenticated account.",
        "operationId": "create_group_api_v1_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Group",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{group_id}": {
      "get": {
        "description": "Return the authenticated group's server-owned settings.",
        "operationId": "get_group_api_v1_groups__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Get Group",
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "description": "Update only settlement policy; authorization remains in GroupService.",
        "operationId": "update_group_api_v1_groups__group_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Update Group",
        "tags": [
          "groups"
        ]
      }
    },
...
  }
}